+++ src/database/source.ts
@@ -4,6 +4,7 @@ import { User } from "./models/User";
 import { UserHistory } from "./models/UserHistory";
 import { ShiritoriChannel } from "./models/ShiritoriChannel";
 import { ShiritoriWord } from "./models/ShiritoriWord";
+import { ShiritoriInflectionRoot } from "./models/ShiritoriInflectionRoot";
 
 import * as dotenv from "dotenv";
 dotenv.config();
@@ -21,7 +22,7 @@ export const PSQLSource = new DataSource({
   database: "sockbot",
   synchronize: false,
   logging: false,
-  entities: [User, UserHistory, ShiritoriChannel, ShiritoriWord],
+  entities: [User, UserHistory, ShiritoriChannel, ShiritoriInflectionRoot, ShiritoriWord],
   migrations: [join(__dirname, "**/migrations/*.{ts,js}")],
   migrationsRun: true,
   subscribers: [],